Add one more filter in zip extraction function

+++ racmo20180917/analysis/Analysis.py
@@ -282,7 +282,7 @@ class Document_Analysis:
                                     f_name = r.filename[:-4]+'_'+str(''.join(fileinzip.split()))
                                     f_name = f_name.replace('/','_')
                                     if r.filename[:-4] in f_name:
-                                        if not os.path.exists(join(PDF_DIR,f_name)):
+                                        if not os.path.exists(join(PDF_DIR,f_name)) and not os.path.exists(join(PDF_DIR,unidecode.unidecode(f_name))):
                                             shutil.move(zfdir, join(PDF_DIR,unidecode.unidecode(f_name)))
                                             z_files.append({"filename":unidecode.unidecode(f_name),"filegroup":r['filegroup'],\
                                                 "filetype":"OTHER",'ext':fileinzip[-3:],"length":len(fileinzip)})
@@ -1180,8 +1180,8 @@ class Document_Analysis:
                                                  creation_date = datetime.datetime.now())
                         model.db.session.add(k)
                         model.db.session.commit()
-                        shutil.copy(join( PDF_DIR,r.filename),join(root_archive,r.filename))
-                        os.remove(join( PDF_DIR,r.filename))
+                        #shutil.copy(join( PDF_DIR,r.filename),join(root_archive,r.filename))
+                        #os.remove(join( PDF_DIR,r.filename))
                 except Exception as e:
                     print(e)
